refactor: move lib/firebase/auth into useAuth
